got basic inference working

+++ vision.py
@@ -6,9 +6,9 @@ import numpy as np
 from typing import Union, Tuple
 from PIL import Image
 
-from .layers import attn, layer_norm, mlp
-from .image_crops import overlap_crop_image
-from .config import VisionConfig
+from layers import attn, layer_norm, mlp
+from image_crops import overlap_crop_image
+from config import VisionConfig
 
 if torch.backends.mps.is_available():
     # Non-divisible input sizes are not implemented on MPS device yet.
@@ -33,7 +33,7 @@ def prepare_crops(
     all_crops = np.transpose(all_crops, (0, 3, 1, 2))
     all_crops = (
         torch.from_numpy(all_crops)
-        .to(device=device, dtype=torch.bfloat16)
+        .to(device=device, dtype=torch.float16)
         .div_(255.0)
         .sub_(0.5)
         .div_(0.5)
